\name{ChaosGamePlot-package}
\alias{ChaosGamePlot-package}
\alias{ChaosGamePlot}
\docType{package}
\title{
  Plot fractals using the chaos game method and C++.
}
\description{
  Plot fractals using the chaos game method and C++.
}
\details{
  This package use only one rule for the chaos game.
}
\author{
Cristian Sandoval-Pineda.

Maintainer: Cristian Sandoval-Pineda <sandovalp.ce@gmail.com>
}
\keyword{ package }
