% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{RunChaosGame}
\alias{RunChaosGame}
\title{RunChaosGame}
\usage{
RunChaosGame(n_vertex, distance, iterations)
}
\arguments{
\item{n_vertex}{The number of vertex for the polygon.}

\item{distance}{The distance between points.}

\item{iterations}{The number of iterations/points for the method.}
}
\value{
A matrix of black/white pixels.
}
\description{
This function allows you to plot some fractals using the chaos game method.
}
\examples{
pixel_matrix = ChaosGamePlot::RunChaosGame(5, .4, 100**3)
}
